anchor change timestamp type

+++ api/red-triplane-scene2d-api/red-triplane-scene2d-api/com/jfixby/r3/ext/api/scene2d/srlz/Anchor.java
@@ -12,14 +12,14 @@ public class Anchor {
 				+ ", position_y=" + position_y + "]";
 	}
 
-	public long time() {
+	public int time() {
 		if (time == null) {
 			return 0;
 		}
 		if ("".equals(time)) {
 			return 0;
 		}
-		return Long.parseLong(time);
+		return (int) Long.parseLong(time);
 	}
 
 }
